requirements.txt
requests
asyncio
python-dotenv
python-telegram-bot
